Very WIP resource table writing.
This works in some but not all cases to create a bit-exact copy of the
original rscs section. However there appear to be many different ways of
writing these. Round-tripping these bit-exact isn't likely possible
without some kind of heuristic to determine the original file's layout.

We now implement this according to the Microsoft spec on ordering of the
tables. But the parser doesn't actually care it seems.